requirements.txt
Pillow
pyserial
pywin32
